Move Timespec impl to common to use in Hermit
